fix(ci): python coverage shard — hosted conformance result metadata is optional
